+++ inlang/source-code/paraglide/paraglide-js-adapter-astro/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @inlang/paraglide-js-adapter-astro
 
+## 0.1.30
+
+### Patch Changes
+
+- @inlang/paraglide-js-adapter-vite@1.2.40
+
+## 0.1.29
+
+### Patch Changes
+
+- @inlang/paraglide-js-adapter-vite@1.2.39
+
 ## 0.1.28
 
 ### Patch Changes