+++ inlang/source-code/paraglide/paraglide-js-adapter-next/examples/app/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @inlang/paraglide-js-adapter-next-example-app
 
+## 0.2.17
+
+### Patch Changes
+
+- Updated dependencies [0774c1a]
+  - @inlang/paraglide-js@1.7.0
+
+## 0.2.16
+
+### Patch Changes
+
+- Updated dependencies [cee4692]
+- Updated dependencies [4b631aa]
+- Updated dependencies [3c7a87c]
+- Updated dependencies [ab1fe48]
+  - @inlang/paraglide-js@1.6.2
+
 ## 0.2.15
 
 ### Patch Changes