+++ inlang/source-code/paraglide/paraglide-js-adapter-vite/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @inlang/paraglide-js-adapter-vite
 
+## 1.2.40
+
+### Patch Changes
+
+- @inlang/paraglide-js-adapter-unplugin@1.4.29
+
+## 1.2.39
+
+### Patch Changes
+
+- @inlang/paraglide-js-adapter-unplugin@1.4.28
+
 ## 1.2.38
 
 ### Patch Changes